Store money as integer minor units to eliminate float rounding
- All amounts (expenses, income, transfers, investments, budgets,
  account opening balances, recurring rules, templates) are now integer
  minor units (paise/cents) instead of double, so arithmetic is exact
  (no more 0.1 + 0.2 drift)
- DB migration v6 converts existing amounts in place (ROUND(x*100))
- Central helpers in currency_format.dart: rupeesToMinor, parseMinor,
  minorToEditString, formatMoney/formatMoneyRounded; inputs convert at
  the UI edge, displays format from minor units
- Providers, account balances, filters, charts, PDF, and CSV updated to
  minor units; charts plot major units; CSV exports human-readable
  major-unit amounts and imports convert back to minor
- Backup format v4 stores minor units; restore auto-converts pre-v4
  (major-unit) backups by ×100 so old backups import correctly
- Tests: v6 migration ×100, conversion helpers incl. the float case,
  and updated existing suites; 57 tests pass

+++ lib/providers/account_provider.dart
@@ -4,7 +4,7 @@ import '../services/db_service.dart';
 
 class AccountProvider extends ChangeNotifier {
   List<Account> _accounts = [];
-  final Map<int, double> _balances = {};
+  final Map<int, int> _balances = {};
 
   List<Account> get accounts => _accounts;
 
@@ -16,7 +16,7 @@ class AccountProvider extends ChangeNotifier {
     return null;
   }
 
-  double? balanceOf(Account account) => _balances[account.id];
+  int? balanceOf(Account account) => _balances[account.id];
 
   Future<void> fetchAccounts() async {
     _accounts = await DBService().getAccounts();
